fix: move dash to config folder
